refactor(plugins): Move deck plugin proxy code to new package
