+++ gate-plugins/src/main/kotlin/com/netflix/spinnaker/gate/plugins/deck/CacheNotReadyException.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.spinnaker.gate.plugins
+package com.netflix.spinnaker.gate.plugins.deck
 
 import com.netflix.spinnaker.kork.exceptions.SystemException
